requirements.txt
Django==1.7
django-annoying==0.8.0
django-debug-toolbar==1.2.1
django-taggit==0.12.1
sqlparse==0.1.11
wsgiref==0.1.2
